+++ pkg/add/add_test.go
@@ -25,7 +25,8 @@ import (
 )
 
 func TestMaskingShouldAddField(t *testing.T) {
-	addMask := NewMask("newvalue")
+	addMask, err := NewMask("newvalue")
+	assert.NoError(t, nil, err, "error should be nil")
 	data := model.NewDictionary().With("field", "SomeInformation")
 	result, err := addMask.MaskContext(data, "newfield")
 	assert.Equal(t, nil, err, "error should be nil")
@@ -36,7 +37,8 @@ func TestMaskingShouldAddField(t *testing.T) {
 func TestFactoryShouldCreateAMask(t *testing.T) {
 	maskingConfig := model.Masking{Selector: model.SelectorType{Jsonpath: "field"}, Mask: model.MaskType{Add: "value"}}
 	mask, present, err := Factory(maskingConfig, 0, nil)
-	waitedMask := NewMask("value")
+	waitedMask, err := NewMask("value")
+	assert.NoError(t, nil, err, "error should be nil")
 	assert.Equal(t, waitedMask, mask, "should be equal")
 	assert.True(t, present, "should be true")
 	assert.Nil(t, err, "error should be nil")